pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "massai-hackathon"
version = "0.1.0"
description = "Skills-Based Hiring Platform API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]==0.110.0",
    "uvicorn==0.27.1",
    "sqlmodel==0.0.14",
    "pydantic==2.6.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.9",
    "python-dotenv==1.0.1",
    "email-validator==2.1.0.post1",
    "nltk>=3.9.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
